requirements.txt
# Web Framework
Flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0

# Machine Learning
tensorflow==2.15.0
numpy==1.26.4
h5py==3.10.0

# Data Processing
pandas==2.2.2
scikit-learn==1.4.2

# Image Processing
Pillow==10.3.0
matplotlib==3.8.4
opencv-python-headless==4.9.0.80

# Backend Services
supabase==2.4.1
python-dotenv==1.0.1